Moved vss to SystemIOWindows. Added GetDirectories to ISystemIO.
