+++ Duplicati/Library/Snapshots/ISystemIO.cs
@@ -57,6 +57,7 @@ namespace Duplicati.Library.Snapshots
         string PathChangeExtension(string path, string extension);
         string PathCombine(string path1, string path2);
         string GetPathRoot(string path);
+        string[] GetDirectories(string path);
         IEnumerable<string> EnumerateFileSystemEntries(string path);
 
         void SetMetadata(string path, Dictionary<string, string> metdata, bool restorePermissions);
